% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{create_sunset_cal}
\alias{create_sunset_cal}
\title{Create a sunset calendar}
\usage{
create_sunset_cal(date = "2014/01/01", lat = 40.7127, long = -74.0059,
  timezone = "America/New_York", num.days = 365, file = "sunset.csv",
  location = "Brooklyn Heights Promenade, Brooklyn, NY 11201")
}
\arguments{
\item{date}{Date at which you want the calendar to start, in yyyy/mm/dd format.}

\item{lat}{Latitude of location (for sunset time calculation)}

\item{long}{Longitude of location (for sunset time calculation, will be negative for continental US)}

\item{timezone}{Timezone of location (for sunset time calculation).}

\item{num.days}{Number of days you want sunset appointments for.}

\item{file}{Filename for outputted .CSV file (to be uploaded to Google Calendar).}

\item{location}{Location of sunset appointment. Will be input into Google Calendar event as the event location.}
}
\description{
This function creates a .CSV of sunset appointments--with a user-specified location--that can be imported into Google Calendar.
}
\examples{
\dontrun{
create_sunset_cal(location = "40.7127, -74.0059")
}
}

